+++ extensions/cornerstone/src/OHIFCornerstoneViewport.js
@@ -1,10 +1,11 @@
 import React, { Component } from 'react';
 
-import OHIFCornerstoneViewportOverlay from './components/OHIFCornerstoneViewportOverlay'
+import OHIFCornerstoneViewportOverlay from './components/OHIFCornerstoneViewportOverlay';
 import ConnectedCornerstoneViewport from './ConnectedCornerstoneViewport';
 import OHIF from '@ohif/core';
 import PropTypes from 'prop-types';
 import cornerstone from 'cornerstone-core';
+import checkForSRAnnotations from './tools/checkForSRAnnotations';
 
 const { StackManager } = OHIF.utils;
 
@@ -15,6 +16,7 @@ class OHIFCornerstoneViewport extends Component {
 
   static defaultProps = {
     customProps: {},
+    isStackPrefetchEnabled: true,
   };
 
   static propTypes = {
@@ -23,6 +25,8 @@ class OHIFCornerstoneViewport extends Component {
     viewportIndex: PropTypes.number,
     children: PropTypes.node,
     customProps: PropTypes.object,
+    stackPrefetch: PropTypes.object,
+    isStackPrefetchEnabled: PropTypes.bool,
   };
 
   static id = 'OHIFCornerstoneViewport';
@@ -184,6 +188,8 @@ class OHIFCornerstoneViewport extends Component {
       displaySet.SOPInstanceUID !== prevDisplaySet.SOPInstanceUID ||
       displaySet.frameIndex !== prevDisplaySet.frameIndex
     ) {
+      const { viewportIndex } = this.props;
+      checkForSRAnnotations({ displaySet, viewportIndex });
       this.setStateFromProps();
     }
   }
@@ -221,7 +227,7 @@ class OHIFCornerstoneViewport extends Component {
       const { displaySet } = this.props.viewportData;
       const { StudyInstanceUID } = displaySet;
 
-      if (currentImageIdIndex > 0) {
+      if (currentImageIdIndex >= 0) {
         this.props.onNewImage({
           StudyInstanceUID,
           SOPInstanceUID: sopInstanceUid,
@@ -231,8 +237,23 @@ class OHIFCornerstoneViewport extends Component {
       }
     };
 
-    const warningsOverlay = props => {
-      return <OHIFCornerstoneViewportOverlay {...props} inconsistencyWarnings={inconsistencyWarnings} />
+    const overlay = props => {
+      const displaySet = this.props.viewportData.displaySet;
+      let filteredSRLabels;
+      if (displaySet.SRLabels && displaySet.SRLabels.length !== 0) {
+        filteredSRLabels = displaySet.SRLabels.filter(
+          SRLabel =>
+            SRLabel.ReferencedSOPInstanceUID === displaySet.SOPInstanceUID
+        );
+      }
+
+      return (
+        <OHIFCornerstoneViewportOverlay
+          {...props}
+          inconsistencyWarnings={inconsistencyWarnings}
+          SRLabels={filteredSRLabels}
+        />
+      );
     };
 
     return (
@@ -243,11 +264,12 @@ class OHIFCornerstoneViewport extends Component {
           imageIdIndex={currentImageIdIndex}
           onNewImageDebounced={newImageHandler}
           onNewImageDebounceTime={300}
-          viewportOverlayComponent={warningsOverlay}
+          viewportOverlayComponent={overlay}
+          stackPrefetch={this.props.stackPrefetch}
+          isStackPrefetchEnabled={this.props.isStackPrefetchEnabled}
           // ~~ Connected (From REDUX)
           // frameRate={frameRate}
           // isPlaying={false}
-          // isStackPrefetchEnabled={true}
           // onElementEnabled={() => {}}
           // setViewportActive{() => {}}
           {...this.props.customProps}